clearing up errors

+++ src/utils.rs
@@ -117,7 +117,7 @@ pub fn chars(arr: [bool; 26]) -> Vec<char> {
         .collect()
 }
 
-pub fn to_word(arr: &Vec<char>) -> Vec<u32> {
+pub fn to_word(arr: &Vec<char>) -> Vec<usize> {
     alph.chars()
         .map(|x| arr.iter().filter(|&y| *y == x).count())
         .collect()
